+++ src/components/cc-invoice/cc-invoice.js
@@ -67,7 +67,7 @@ export class CcInvoice extends LitElement {
         ${!this.error ? html`
           <div slot="button">${ccLink(invoice.downloadUrl, i18n('cc-invoice.download-pdf'), skeleton)}</div>
           <div class="info"><em class=${classMap({ skeleton })}>${i18n('cc-invoice.info', { date, amount })}</em></div>
-          <cc-html-frame class="frame" ?loading="${skeleton}" title="${i18n('cc-invoice.title')} ${number}">
+          <cc-html-frame class="frame" ?loading="${skeleton}" iframe-title="${i18n('cc-invoice.title')} ${number}">
             ${!skeleton ? html`
               <template>${unsafeHTML(this.invoice.invoiceHtml)}</template>
             ` : ''}